feat: implement post method for vote model

+++ DevTestApi/src/main/java/salesiana/university/dev/test/api/mappers/requests/vote/CreateVoteRequest.java
@@ -3,7 +3,7 @@ package salesiana.university.dev.test.api.mappers.requests.vote;
 import jakarta.validation.constraints.Email;
 
 public record CreateVoteRequest(
-        int optionId,
+        Long optionId,
 
         @Email(message = "VoterEmail must be a valid email address.")
         String voterEmail
